requirements_retriever.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
langgraph>=0.0.20
psycopg2-binary>=2.9.0
pgvector>=0.2.0
google-generativeai>=0.3.0
duckduckgo-search>=3.9.0
pypdf>=3.0.0
python-docx>=0.8.11
numpy>=1.24.0
redis>=4.5.0
pandas>=1.5.0
